+++ RELEASE-NOTES.md
@@ -39,6 +39,8 @@ The project release numbers follow [Semantic Versioning](http://semver.org/spec/
 - HIP CMake utilities updated to AMD's latest version
 - Updated ``add_code_coverage_target`` to ``blt_add_code_coverage_target``, which now supports
   user-specified source directories
+- Code coverage targets leave LCOV-generated files intact for later use; these files will
+  still be removed by ``make clean``
 
 ### Fixed
 - ClangFormat checks now support multiple Python executable names
@@ -61,6 +63,8 @@ The project release numbers follow [Semantic Versioning](http://semver.org/spec/
 - ``blt_patch_target`` no longer attempts to set system include directories when a target
   has no include directories
 - Header-only libraries now can have dependencies via DEPENDS_ON in ``blt_add_library``
+- Added a workaround for include directories of imported targets on PGI. CMake was 
+  erroneously marking them as SYSTEM but this is not supported by PGI.
 
 ## [Version 0.3.6] - Release date 2020-07-27